requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
future==0.17.1
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pylint==2.3.1
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
